Move Plotly code to JS file
